+++ py-polars/polars/__init__.py
@@ -66,7 +66,7 @@ from polars.exceptions import (
     StructFieldNotFoundError,
 )
 from polars.expr.expr import Expr
-from polars.internals.functions import (
+from polars.functions.eager import (
     align_frames,
     concat,
     cut,
@@ -75,8 +75,7 @@ from polars.internals.functions import (
     ones,
     zeros,
 )
-from polars.internals.io import read_ipc_schema, read_parquet_schema
-from polars.internals.lazy_functions import (
+from polars.functions.lazy import (
     all,
     any,
     apply,
@@ -125,10 +124,11 @@ from polars.internals.lazy_functions import (
     tail,
     var,
 )
-from polars.internals.lazy_functions import date_ as date
-from polars.internals.lazy_functions import datetime_ as datetime
-from polars.internals.lazy_functions import list_ as list
-from polars.internals.whenthen import when
+from polars.functions.lazy import date_ as date
+from polars.functions.lazy import datetime_ as datetime
+from polars.functions.lazy import list_ as list
+from polars.functions.whenthen import when
+from polars.internals.io import read_ipc_schema, read_parquet_schema
 from polars.io import (
     read_avro,
     read_csv,
