+++ py-polars/polars/internals/type_aliases.py
@@ -15,7 +15,7 @@ if TYPE_CHECKING:
     from polars.dependencies import pandas as pd
     from polars.dependencies import pyarrow as pa
     from polars.expr.expr import Expr
-    from polars.internals.whenthen import WhenThen, WhenThenThen
+    from polars.functions.whenthen import WhenThen, WhenThenThen
     from polars.series.series import Series
 
     if sys.version_info >= (3, 10):
