+++ py-polars/polars/lazyframe/groupby.py
@@ -368,7 +368,7 @@ class LazyGroupBy(Generic[LDF]):
         └────────┴───────┘
 
         """
-        return self.agg(pli.lazy_functions.count())
+        return self.agg(pli.count())
 
     def first(self) -> LDF:
         """
